Add timeout to Swift Recon API Call
When a node is unavailable, the service would currently wait forever for
the Swift Recon API call to return. Eventually the client would timeout
and this would cause a "[Errno 32] Broken Pipe" error causing the
service to crash. Add a configurable timeout to help prevent an offline
node causing the service to stop functioning.